feat(strategy): Introduce strict-github strategy
This strategy is equal to the GitHub strategy except it will throw
an error if the first commit title is not equal to the pull request
title. Regarding the GitHub flow, the squash commit title depends on
the number of the commit messages in the feature branch. This strict
mode makes the behavior consistent and ensures the title represent
actual squash commit info